+++ lizard_portal/templates/portals/analyse-interpretatie.js
@@ -30,8 +30,10 @@
                   width: 800,
                   height: 600,
                   modal: true,
+                  constraint: true,
                   finish_edit_function: function (updated_record) {
-                      //todo
+                      var store = Ext.StoreManager.lookup('analyse_store');
+                      store.load({params: {object_ident: Lizard.CM.getContext().object.id}});
                   },
                   editpopup: true,
                   loader:{