+++ lizard_portal/templates/portals/maatregelen.js
@@ -12,7 +12,7 @@
 	items:[{
 		flex: 1,
 		items: [{
-            title: 'Maatregelen voor' + Lizard.CM.context.object.name + ' grafiek',
+            title: 'Maatregelengrafiek',
             flex: 1,
             xtype: 'multigraphstore',
             plugins: [
@@ -39,7 +39,7 @@
                 }]
             })
         },{
-			title: 'Maatregelen voor ' + Lizard.CM.context.object.name,
+			title: 'Maatregelen',
             flex:2,
             autoScroll: true,
             plugins: [
@@ -76,6 +76,7 @@
                         height: 600,
                         modal: true,
                         editpopup: true,
+                        constrain: true,
                         loader:{
                             loadMask: true,
                             autoLoad: true,